# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: WorkTime Sync API
  version: 0.1.0
  description: |
    Веб-платформа для актуализации рабочего времени сотрудников.
    Endpoints разделены на публичные (`/auth/*`, `/webhooks/*`) и защищённые JWT (всё остальное).
  contact:
    name: WorkTime Sync team
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Создать пользователя и сразу получить токены
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Получить access+refresh токены
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Обменять refresh-токен на новую пару
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh: { type: string }
      responses:
        '200': { description: ok }
  /api/v1/me:
    get:
      tags: [me]
      security: [{ bearerAuth: [] }]
      summary: Полная карточка текущего пользователя
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
...
  /api/v1/exceptions:
    get:
      tags: [exceptions]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200': { description: ok }
    post:
      tags: [exceptions]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, start_at, end_at]
              properties:
                kind: { type: string, enum: [vacation, sick_leave, business_trip, personal_hours, custom] }
                start_at: { type: string, format: date-time }
                end_at: { type: string, format: date-time }
                comment: { type: string }
      responses:
        '201': { description: created }
...
  /api/v1/teams:
    get:
      tags: [teams]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: ok }
...
  /api/v1/employees:
    get:
      tags: [employees]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: ok }
  /api/v1/employees/{id}:
    get:
      tags: [employees]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
  /api/v1/hr/roadmap:
    get:
      tags: [hr]
      security: [{ bearerAuth: [] }]
      summary: Приоритизированная дорожная карта HR
      responses:
        '200': { description: ok }
...
  /api/v1/admin/users:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Список пользователей (только admin)
      responses:
        '200': { description: ok }
  /api/v1/admin/sources:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Все интеграции в системе
      responses:
        '200': { description: ok }
  /api/v1/admin/rules:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Текущие веса риска (w1..w5)
      responses:
        '200': { description: ok }
    put:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Обновить веса (сумма должна быть = 1.0)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [w1, w2, w3, w4, w5, freshness_d_days]
              properties:
                w1: { type: number }
                w2: { type: number }
                w3: { type: number }
                w4: { type: number }
                w5: { type: number }
                freshness_d_days: { type: integer }
      responses:
        '204': { description: ok }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        full_name: { type: string }
        role: { type: string, enum: [admin, employee, manager, hr, pm, analyst] }
        timezone: { type: string, example: Europe/Moscow }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        tokens:
          type: object
          properties:
            access: { type: string }
            refresh: { type: string }
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { type: string }
        full_name: { type: string }
        timezone: { type: string }
    MeResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        employee: { type: object }
        work_profile: { type: object }
        exceptions: { type: array, items: { type: object } }
...